// backend/src/cliente-cartera/dto/create-contacto-cliente.dto.ts
export class CreateContactoClienteDto {
  nombres!: string;
  apellidos?: string;
  telefono?: string;
  email?: string;
  cargo?: string;
  etapa?: string;
  source?: string;
  clienteRecuperado?: string;
  departamento?: string;
  provincia?: string;
  distrito?: string;
  direccion?: string;
  assignedTo?: string;
  clienteEmpresaId?: string;
  isPrimary?: boolean;
}