// backend/src/cliente-cartera/dto/update-contacto-cliente.dto.ts
export class UpdateContactoClienteDto {
  nombres?: string;
  apellidos?: string;
  telefono?: string;
  email?: string;
  cargo?: string;
  etapa?: string;
  source?: string;
  clienteRecuperado?: string;
  departamento?: string;
  provincia?: string;
  distrito?: string;
  direccion?: string;
  assignedTo?: string;
}